+++ pkg/spark/dialer.go
