+++ pkg/spark/doc.go
